Refactor: Refatorando os scripts de gráficos
