+++ index.html
@@ -186,7 +186,7 @@
     <!-- <script src="./src/loadTransactions.js"></script> -->
     <!-- <script src="./src/account.js"></script> -->
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-    <script type="module" src="./src/script.js"></script>
+    <script type="module" src="./src/script.js" defer></script>
 </body>
 
 </html>
